all mitigations complete - Tanish H
